changed Structure.colors to being a class attribute, GeomCollection is a child of SuperStructure, small bug fix and code cleanup
